+++ src/models/spaces/populations/mouse-model/mouse-populations-model.ts
@@ -385,7 +385,8 @@ export const MousePopulationsModel = types
         destroyInteractive() {
           interactive = undefined;
           setupGraph();
-        }
+        },
+        setupGraph
       }
     };
   })